requirements.txt
openai
flask
streamlit
requests
python-dotenv
pandas
ollama
